Node colors
Whenever we make changes with node.data(), node.addClass(), etc,
we need to force a refresh of components that depend on Cytoscape.

This is because this.state.cytoscape is just a reference to
the Cytoscape object, which the Cytoscape API modifies directly.

By convention, in this situation we will call:

this.setState({
    cytoscape: this.state.cytoscape
});